refactor(all_orders_for_member): adjust after moving the DTOs in new packages
