+++ src/main/kotlin/bass/controller/admin/AdminController.kt
@@ -4,8 +4,8 @@ import bass.annotation.CheckAdminOnly
 import bass.annotation.IgnoreCheckLogin
 import bass.controller.admin.usecase.FindMembersWithRecentCartActivityUseCase
 import bass.controller.admin.usecase.FindTopProductsUseCase
-import bass.dto.product.TopProductDTO
 import bass.dto.member.ActiveMemberDTO
+import bass.dto.product.TopProductDTO
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
