+++ src/main/kotlin/bass/controller/order/usecase/OrderCreationUseCase.kt
@@ -1,12 +1,19 @@
 package bass.controller.order.usecase
 
-import bass.dto.order.OrderDTO
 import bass.dto.member.MemberLoginDTO
+import bass.dto.order.OrderDTO
+import bass.dto.order.OrderResponseDTO
 import bass.model.PaymentRequest
+import org.springframework.data.domain.Pageable
 
 interface OrderCreationUseCase {
     fun create(
         memberLoginDTO: MemberLoginDTO,
         paymentRequest: PaymentRequest,
     ): OrderDTO
+
+    fun getOrdersByMemberId(
+        memberId: Long,
+        pageable: Pageable,
+    ): List<OrderResponseDTO>
 }
