+++ src/main/kotlin/bass/dto/TopProductDTO.kt
@@ -1,4 +1,3 @@
 package bass.dto.product
 
 // stays as OffsetDateTime because Instant causes a conflict with the result from native query
-
