+++ src/main/kotlin/bass/dto/naturalSearch/NaturalSearchResponseDTO.kt
@@ -1,7 +1,7 @@
 package bass.dto.naturalSearch
 
-import bass.dto.tag.TagInferenceResultDTO
 import bass.dto.meal.MealResponseDTO
+import bass.dto.tag.TagInferenceResultDTO
 import bass.entities.MealEntity
 import bass.mappers.toDTO
