+++ src/main/kotlin/bass/services/meal/MealServiceImpl.kt
@@ -2,11 +2,11 @@ package bass.services.meal
 
 import bass.controller.meal.usecase.AISearchUseCase
 import bass.controller.meal.usecase.CrudMealUseCase
-import bass.dto.naturalSearch.NaturalSearchRequestDTO
-import bass.dto.naturalSearch.NaturalSearchResponseDTO
 import bass.dto.meal.MealPatchDTO
 import bass.dto.meal.MealRequestDTO
 import bass.dto.meal.MealResponseDTO
+import bass.dto.naturalSearch.NaturalSearchRequestDTO
+import bass.dto.naturalSearch.NaturalSearchResponseDTO
 import bass.entities.MealEntity
 import bass.exception.InvalidTagNameException
 import bass.exception.NotFoundException
