+++ src/test/kotlin/bass/endtoend/CartItemE2ETest.kt
@@ -2,8 +2,8 @@ package bass.endtoend
 
 import bass.dto.cartItem.CartItemRequestDTO
 import bass.dto.cartItem.CartItemResponseDTO
-import bass.dto.page.PageResponseDTO
 import bass.dto.meal.MealResponseDTO
+import bass.dto.page.PageResponseDTO
 import io.restassured.RestAssured
 import io.restassured.common.mapper.TypeRef
 import io.restassured.http.ContentType
