+++ src/test/kotlin/bass/integration/OrderServiceTest.kt
@@ -1,7 +1,7 @@
 package bass.integration
 
-import bass.dto.order.OrderDTO
 import bass.dto.member.MemberLoginDTO
+import bass.dto.order.OrderDTO
 import bass.entities.AchievementEntity
 import bass.entities.CartItemEntity
 import bass.entities.MealEntity
